refactor Tones into component
